By default, this project (parent) clones dependencies (children) recursively (their children and so on).
By non-default, this project allows itself to be built when when its children and so on, are pre-built.
This is how the meta-aws project handles dependencies.

In the non-default case, the sources are not compiled which infers that the sources are not in the
source tree - i.e. -DBUILD_DEPS=OFF.

The CMakeFiles assume that the dependent sources are in the source tree.

This commit allows building this project when the sources are not in the source tree by adding guards
around the statements requiring the dependent sources to be in the source tree.