Use latest versions of S2N-TLS and AWS-LC (#470)
Update all submodules, including S2N-TLS and AWS-LC which we hadn't updated for a while due to AWS-LC changing the default visibility of symbols in static builds from "hidden" to "visible" (see https://github.com/awslabs/aws-lc/pull/466)

**ISSUE**
This resulted in the symbols being visible in whatever application ultimately used libcrypto.a. If an application ALSO ended up using libcrypto.so from the system's OpenSSL installation, we'd get crazy crashes due to SOME libcrypto function calls routing through libcrypto.a and SOME function calls routing through libcrypto.so.

**DESCRIPTION OF CHANGES:**
Tweak the pkcs11_connect sample so it explicitly hides symbols from libcrypto.a. This sample is likely to load the libcrypto.so from they system's OpenSSL installation as a side-effect of loading libsofthsm2.so.

If your application starts crashing after taking this change, with a stacktrace showing math calls to libcrypto, you'll need to do the same.